requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
python-dotenv==1.0.1
requests==2.31.0
flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
